feat: add harden-one-flatpak ujust command
adds a new ujust command to apply the hardened_malloc envar to a specified flatpak. This way, the audit-ujust script can recommend users run `ujust harden-one-flatpak $someflatpak` to enable the best hwcap version of their hardened_malloc library as an override for a particular flatpak (as opposed to the global override, which specific overrides take priority over).